requirements.txt
-r requirements_dev.txt
colorlog>=6.10.1
hassil>=3.5.0
home_assistant_intents>=2026.3.0
pytest-homeassistant-custom-component>=0.13.317
python-telegram-bot[socks]==22.1
python-typing-update>=0.6.0
pip>=21.3.1
pre-commit>=4.5.1
pre-commit-hooks>=6.0.0
codespell>=2.4.1
ruff>=0.15.1
yamllint>=1.37.1
zizmor>=1.22.0
telegramify_markdown[mermaid]==1.1.1
PyTurboJPEG>=1.8.0
